Fixed some compilation bugs

+++ src/Field.cpp
@@ -47,9 +47,9 @@ const T& Field<T>::at(int i, int j, int k) const {
 
 template<class T>
 const T Field<T>::at(double x, double y, double z) const {
-    int i = (int)floor(x/dx);
-    int j  =(int)ceil(y/dy);
-    int k = (int)floor(z/dz);
+    int i = floor(x/dx);
+    int j = ceil(y/dy);
+    int k = floor(z/dz);
 
     double lambda = x/dx - (i + 1);
     double mu = z/dz - (k + 1);
